Remove the now useless AbstractFunctionCallDefinition
Merge it in MethodDefinition, this is much simpler and means one less file to load.
